+++ api/src/routes.ts
@@ -7,6 +7,8 @@ import {
   searchUsernames,
   getLatestChatMessages,
   getOlderChatMessages,
+  sendExampleTweet,
+  examlpeUpdateTweetsPerformance,
 } from '@/controller';
 
 export const router = express.Router();
@@ -16,6 +18,8 @@ router.get('/', getUsernames);
 router.get('/refresh/:id', refreshScore);
 router.get('/single/:id', getUsernameById);
 router.get('/search', searchUsernames);
+router.get('/example-tweet', sendExampleTweet);
+router.get('/example-update-tweets-performance', examlpeUpdateTweetsPerformance);
 
 router.get('/chat/messages', getLatestChatMessages);
 router.get('/chat/messages/:oldestMessageId', getOlderChatMessages);